+++ HISTORY.md
@@ -1,3 +1,8 @@
+- wspot v1.0.3 changelog:
+1. [exploit-1:] fixed issues with preparing stage.
+2. [exploit-1:] improved error detection.
+3. [wspot:] improved error notification.
+4. [iproute-parser:] added ssid optimizer.
 - wspot v1.0.2 changelog:
 1. [installer:] fixed auto updater.
 2. [poster:] fixed latest tag release upload.